pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datamil"
version = "0.1.0"
description = "DataMIL: Selecting Data for Robot Imitation Learning with Datamodels"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Shivin Dass", email = "shivindass@gmail.com"},
    {name = "Alaa Khaddaj"},
    {name = "Logan Engstrom"},
    {name = "Aleksander Madry"},
    {name = "Andrew Ilyas"},
    {name = "Roberto Martín-Martín"},
]
keywords = ["robotics", "imitation-learning", "machine-learning", "data-selection"]
requires-python = ">=3.10"
dependencies = [
    "gym >= 0.26",
    "numpy == 1.24.3",
    "ml_dtypes == 0.2.0",
    "chex == 0.1.85",
    "optax == 0.1.5",
    "tensorflow_probability == 0.23.0",
    "tensorflow == 2.15.0",
    "jax == 0.4.20",
    "distrax == 0.1.5",
    "flax == 0.7.5",
    "ml_collections >= 0.1.0",
    "tqdm >= 4.60.0",
    "absl-py >= 0.12.0",
    "scipy == 1.12.0",
    "wandb >= 0.12.14",
    "einops >= 0.6.1",
    "imageio >= 2.31.1",
    "moviepy >= 1.0.3",
    "transformers >= 4.34.1",
    "tensorflow_hub >= 0.14.0",
    "tensorflow_text >= 2.13.0",
    "tensorflow_datasets == 4.9.2",
    "tensorflow_graphics == 2021.12.3",
    "dlimp @ git+https://github.com/kvablack/dlimp@d08da3852c149548aaa8551186d619d87375df08",
    "plotly >= 5.16.1",
    "matplotlib",
    "flatten-dict==0.4.2",
    "dill==0.3.9",
    "cupy-cuda11x==13.3.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit == 3.3.3",
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]

[project.urls]
Homepage = "https://github.com/ShivinDass/datamil"
Repository = "https://github.com/ShivinDass/datamil"
Documentation = "https://github.com/ShivinDass/datamil#readme"
"Bug Tracker" = "https://github.com/ShivinDass/datamil/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["datamil"]

[tool.hatch.build.targets.sdist]
include = [
    "/datamil",
    "/README.md",
    "/LICENSE",
]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=datamil --cov-report=html --cov-report=term-missing"
